tutorials/object-storage-django-storages/example-project/requirements.txt
Django
django-environ
django-storages[boto3]
whitenoise[brotli]